requirements.txt
python-dotenv==1.0.1
langchain==0.2.14
langchain-community==0.2.12
langchain-openai==0.1.22
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
transformers==4.42.3
accelerate==0.33.0
torch>=2.2.0
pypdf==4.2.0
unstructured==0.14.9
beautifulsoup4==4.12.3
requests==2.32.3
streamlit==1.36.0
rank-bm25==0.2.2
groq==0.9.0
